fix: upload Cover Image file error for Safari Browser
ShowImage boolean was being set to true before the file is read. This triggered the OnError() callback of the img tag in Safari.

For some reason Google Chrome, does not complain about this.

After removing the extra setShowImage(true) line, upload cover image using a file works for both Chrome and Safari.